feat(resume): Create Resume UI, Install Moment
Numerous changes in order to flesh out the Resume view.

Updated rating and svg-loader to properly load the rating image and apply styles to it dynamically.

Installed Moment and Ignore-Loader dependencies. Ignore-Loader is used specifically to exclude specific file types from the Webpack bundling process.

Removed base64.css, moved it's code into spash.css. This was a holdover that is no longer necessaary.

Temporarily added overflow scrolling to the router-view. This will later be replaced with a better solution, and router-view will be hidden from the DOM with the containerless attribute.

Added new fonts.

Added new rating.svg