fix(nav): Replace GitHub GraphQL API w/ Rest API, Fix GH-Pages Redirect
Added github-api dependency to handle interaction with GitHub's public API.

Fixed an issue with app-nav not being able to pull data via GitHub's graphql api because it is not currently available to the public. Replaced the api call with a standard rest call via github-api library.

Added a redirect hack via 404.html as outlined in this blog post:
http://www.backalleycoder.com/2016/05/13/sghpa-the-single-page-app-hack-for-github-pages/

This fixes an issue with gh-pages not supporting redirection to index.html for static single-page apps.

Updated webpack config to include 404.html when bundling.

Added a new route to app.js for galleries.