feat(nav): Add Social Media Links, Remove Incomplete Routes
Updated the app navigation with social media links which are defined in app.config. Also temporarily set all incomplete sections of the main site navigation to nav: false in app.js.

Began reformatting app-nav.scss to be responsive.

Added an under construction filler content to the Blog route because it is linked on social media profiles. Resume is currently the default route for the application.

Moved some redundant CSS for the main element to global.scss and removed it from individual routes to keep the stylesheets DRY.