chore(dependencies): Update Project Dependencies, Resolve Build Errors
Updated project dependencies to the latest versions and resolved webpack build errors that cropped up as a result.

There seems to be an issue with Uglifyjs and object property names, so a small change needed to be made to store.js to resolve this issue. This can likely be reverted later on.