feat(mobile): Fix Mobile Webkit Specific Styling Errors
Mobile Safari (and by extension Chrome and Firefox on iOS) has some weird rules for the flexbox specification, making it so certain elements (such as button) cannot be display: flex. The solution is to wrap all button content in a span element. Similarly there are issues with absolute positioning and flex items, again using a span element as a wrapper can solve for inconsistencies.

Added initial support for Webpack's Offline plugin.